chore: rename EmailConnectorV2. Remove unused commented code
