+++ packages/wagmi/exports/react.tsx
@@ -32,7 +32,4 @@ export {
 } from '@web3modal/scaffold-react'
 
 // -- Universal Exports -------------------------------------------------------
-
-// A export { EIP6963Connector } from '../src/connectors/EIP6963Connector.js'
-
 export { defaultWagmiConfig } from '../src/utils/defaultWagmiReactConfig.js'
