+++ packages/wagmi/exports/vue.ts
@@ -30,6 +30,4 @@ export {
 } from '@web3modal/scaffold-vue'
 
 // -- Universal Exports -------------------------------------------------------
-
-//A export { EIP6963Connector } from '../src/connectors/EIP6963Connector.js'
 export { defaultWagmiConfig } from '../src/utils/defaultWagmiCoreConfig.js'
